feat: add base styling, style search
