+++ src/features/SearchForm/ui/SearchForm.module.css
@@ -1,0 +1,21 @@
+.form {
+  position: sticky;
+  top: 0;
+  z-index: 10;
+  display: flex;
+  gap: 1em;
+  border-radius: 0 0 var(--border-radius) var(--border-radius);
+  padding: 1.25rem 3rem;
+  box-shadow: var(--shadow-md);
+  color: var(--text-on-red);
+  font-weight: 700;
+  font-size: 1.5rem;
+  background: var(--bg-header);
+}
+
+@media (max-width: 768px) {
+  .form {
+    gap: 0.5em;
+    border-radius: 0;
+  }
+}
