+++ src/shared/styles/index.ts
@@ -1,0 +1,3 @@
+import './index.css';
+import './variables.css';
+import './normalize.css';
