+++ src/shared/ui/Loader/Loader.tsx
@@ -1,0 +1,17 @@
+import { Component } from 'react';
+import styles from './Loader.module.css';
+
+interface LoaderProps {
+  className?: string;
+}
+
+export class Loader extends Component<LoaderProps> {
+  render() {
+    const { className = '' } = this.props;
+    return (
+      <div className={`${styles['spinner-container']} ${styles[className]}`}>
+        <div className={styles['spinner-ring']} />
+      </div>
+    );
+  }
+}
